Move integration tests to separate directory
Again, we do all of this to prepare for creating "real" test suites for
each type of tests.
