+++ .gitignore
@@ -4,6 +4,6 @@ composer.phar
 node_modules
 .DS_Store
 Thumbs.db
-/tests/tmp
+/tests/integration/tmp
 .vagrant
 .idea/*
