+++ tests/integration/api/Controller/CreateDiscussionControllerTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Flarum\Tests\Api\Controller;
+namespace Flarum\Tests\integration\api\Controller;
 
 use Flarum\Api\Controller\CreateDiscussionController;
 use Flarum\Discussion\Discussion;
