+++ tests/integration/installer/DefaultInstallationTest.php
